logging/requirements.txt
# Python dependencies for Multi-Container Monitoring Log API
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
